// apps/web/app/api/student/models.ts
export type AddStudentModel = {
  aadharCardNumber: string;
  admissionMode: string;
  admissionType: string;
  annualIncome: string;
  bloodGroup: string;
  caste: string;
  community: string;
  dob: string;
  emailId: string;
  fatherName: string;
  fatherOccupation: string;
  firstLanguage: string;
  firstName: string;
  guardiansOccupation: string;
  gender: string;
  guardianName: string;
  guardianPhoneNumber: string;
  lastName: string;
  maritalStatus: string;
  mediumOfEducation10th: string;
  mediumOfEducation12th: string;
  middleName: string;
  mobileNumber: string;
  motherName: string;
  motherOccupation: string;
  motherTongue: string;
  nationality: string;
  noOfSiblings: string;
  obtainedMark10th: string;
  obtainedMark12th: string;
  courseOption1: string;
  courseOption2: string;
  parentsSeparated: string;
  permanentDistrict: string;
  permanentPostalCode: string;
  permanentState: string;
  phoneNumber: string;
  relationshipType: string;
  religion: string;
  residentialDistrict: string;
  residentialPostalCode: string;
  residentialState: string;
  scholarship: string;
  schoolName10th: string;
  schoolName12th: string;
  siblingClass1: string;
  siblingClass2: string;
  siblingName1: string;
  siblingName2: string;
  siblingRelation1: string;
  siblingRelation2: string;
  yearOfPassing10th: string;
  yearOfPassing12th: string;
  residentialAddress: string;
  permanentAddress: string;
};